+++ packages/user-service/CHANGELOG.md
@@ -4,6 +4,7 @@
 
 - refactor: getUserInfo from command in O2
 - fix: set timeout for getUserInfoFromDefClient
+- fix: only run getUserInfoFromDefClient in AliInternal
 
 ## 0.1.0